Message edition

+++ Stripchat/Libs/Helper.h
@@ -14,7 +14,7 @@
 + (void)showWarningWithMsg:(NSString *)msg delegate:(id<UIAlertViewDelegate>)delegate;
 + (void)showErrorWithMsg:(NSString *)msg delegate:(id<UIAlertViewDelegate>)delegate;
 + (void)showValidationWithMsg:(NSString *)msg delegate:(id<UIAlertViewDelegate>)delegate;
-+ (NSString *)getImageWithUrl:(NSString *)url witdh:(NSUInteger)width height:(NSUInteger)height;
++ (NSString *)getImageWithUrl:(NSString *)url witdh:(CGFloat)width height:(CGFloat)height;
 + (BOOL)isIPhoneDevice;
 
 @end
